Moved non-game classes to util package.
Also moved FastNoiseLite to an "fnl" package but I'm not quite happy about
it. The alternative would be to leave it under src, but I feel it's not an
appropriate place. Thing is, Haxe seems to enforce
directories == packages, so I can't both organize files while keeping
a decoupled package structure.